+++ app/include/tab/setting_tab.hpp
@@ -48,10 +48,13 @@ private:
     BRLS_BIND(brls::BooleanCell, btnTouchGesture, "setting/player/touch_gesture");
     BRLS_BIND(brls::BooleanCell, btnClipPoint, "setting/player/clip_point");
     BRLS_BIND(brls::BooleanCell, btnFullscreen, "setting/fullscreen");
+    BRLS_BIND(brls::BooleanCell, btnAlwaysOnTop, "setting/always_on_top");
+    BRLS_BIND(brls::BooleanCell, btnSingle, "setting/single");
     BRLS_BIND(brls::BooleanCell, btnOverClock, "setting/overclock");
     BRLS_BIND(brls::BooleanCell, btnDebug, "setting/debug");
     BRLS_BIND(brls::SelectorCell, inputThreads, "setting/network/threads");
     BRLS_BIND(brls::SelectorCell, selectorTimeout, "setting/network/timeout");
+    BRLS_BIND(brls::BooleanCell, btnTls, "setting/network/tls");
     BRLS_BIND(brls::BooleanCell, btnProxy, "setting/network/proxy");
     BRLS_BIND(brls::InputCell, inputProxyHost, "setting/network/proxy_host");
     BRLS_BIND(brls::InputNumericCell, inputProxyPort, "setting/network/proxy_port");