+++ app/src/utils/misc.cpp
@@ -4,6 +4,13 @@
 #include <sstream>
 #include <iomanip>
 
+#if defined(__APPLE__) || defined(__linux__)
+#include <unistd.h>
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <sys/un.h>
+#endif
+
 #if defined(_WIN32) && !defined(_WINRT_)
 
 #define WIN32_LEAN_AND_MEAN
@@ -35,9 +42,12 @@ LONG WINAPI createMiniDump(_EXCEPTION_POINTERS* pep) {
     auto fnSymInitialize = (WINBOOL(WINAPI*)(HANDLE, PCSTR, WINBOOL))::GetProcAddress(dbghelp, "SymInitialize");
     auto fnSymCleanup = (WINBOOL(WINAPI*)(HANDLE))::GetProcAddress(dbghelp, "SymCleanup");
     auto fnStackWalk64 = (WINBOOL(WINAPI*)(DWORD, HANDLE, HANDLE, LPSTACKFRAME64, PVOID, PREAD_PROCESS_MEMORY_ROUTINE64,
-        PFUNCTION_TABLE_ACCESS_ROUTINE64, PGET_MODULE_BASE_ROUTINE64, PTRANSLATE_ADDRESS_ROUTINE64))::GetProcAddress(dbghelp, "StackWalk64");
-    auto fnSymGetSymFromAddr64 = (WINBOOL(WINAPI*)(HANDLE, DWORD64, PDWORD64, PIMAGEHLP_SYMBOL64))::GetProcAddress(dbghelp, "SymGetSymFromAddr64");
-    auto fnSymGetModuleInfo64 = (WINBOOL(WINAPI*)(HANDLE, DWORD64, PIMAGEHLP_MODULE64))::GetProcAddress(dbghelp, "SymGetModuleInfo64");
+        PFUNCTION_TABLE_ACCESS_ROUTINE64, PGET_MODULE_BASE_ROUTINE64,
+        PTRANSLATE_ADDRESS_ROUTINE64))::GetProcAddress(dbghelp, "StackWalk64");
+    auto fnSymGetSymFromAddr64 = (WINBOOL(WINAPI*)(HANDLE, DWORD64, PDWORD64, PIMAGEHLP_SYMBOL64))::GetProcAddress(
+        dbghelp, "SymGetSymFromAddr64");
+    auto fnSymGetModuleInfo64 =
+        (WINBOOL(WINAPI*)(HANDLE, DWORD64, PIMAGEHLP_MODULE64))::GetProcAddress(dbghelp, "SymGetModuleInfo64");
     auto fnSymFTA64 = (PFUNCTION_TABLE_ACCESS_ROUTINE64)::GetProcAddress(dbghelp, "SymFunctionTableAccess64");
     auto fnSymGMB64 = (PGET_MODULE_BASE_ROUTINE64)::GetProcAddress(dbghelp, "SymGetModuleBase64");
 
@@ -97,9 +107,9 @@ LONG WINAPI createMiniDump(_EXCEPTION_POINTERS* pep) {
     CHAR tempPath[MAX_PATH];
     DWORD cbTemp = ::GetTempPathA(sizeof(tempPath), tempPath);
     ::GetLocalTime(&lt);
-    
-    snprintf(tempPath + cbTemp, sizeof(tempPath) - cbTemp, "switchfin-%04d%02d%02d-%02d%02d%02d.dmp", 
-        lt.wYear, lt.wMonth, lt.wDay, lt.wHour, lt.wMinute, lt.wSecond);
+
+    snprintf(tempPath + cbTemp, sizeof(tempPath) - cbTemp, "switchfin-%04d%02d%02d-%02d%02d%02d.dmp", lt.wYear,
+        lt.wMonth, lt.wDay, lt.wHour, lt.wMinute, lt.wSecond);
     HANDLE hDump = ::CreateFileA(tempPath, GENERIC_WRITE, 0, nullptr, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL, nullptr);
 
     MINIDUMP_EXCEPTION_INFORMATION info;
@@ -145,4 +155,39 @@ std::string misc::hexEncode(const unsigned char* data, size_t len) {
         ss << std::hex << std::setw(2) << std::setfill('0') << (int)data[i];
     }
     return ss.str();
+}
+
+bool misc::sendIPC(const std::string& sock, const std::string& payload) {
+#ifdef _WIN32
+    HANDLE hPipe = CreateFile(sock.c_str(), GENERIC_WRITE, 0, NULL, OPEN_EXISTING, 0, NULL);
+    if (hPipe == INVALID_HANDLE_VALUE) return false;
+
+    if (!WriteFile(hPipe, payload.c_str(), payload.size(), NULL, NULL)) {
+        brls::Logger::warning("sendIPC `{}` failed: {}", payload, GetLastError());
+    }
+
+    CloseHandle(hPipe);
+    return true;
+#elif defined(__APPLE__) || defined(__linux__)
+    int fd = socket(AF_UNIX, SOCK_STREAM, 0);
+    if (fd == -1) return false;
+
+    struct sockaddr_un addr;
+    addr.sun_family = AF_UNIX;
+    strncpy(addr.sun_path, sock.c_str(), sizeof(addr.sun_path) - 1);
+
+    if ((connect(fd, (struct sockaddr*)&addr, sizeof(addr))) == -1) {
+        close(fd);
+        return false;
+    }
+
+    if (write(fd, payload.c_str(), payload.size()) == -1) {
+        brls::Logger::warning("sendIPC `{}` failed: {}", payload, errno);
+    }
+
+    close(fd);
+    return true;
+#else
+    return false;
+#endif
 }