+++ app/src/view/player_setting.cpp
@@ -98,15 +98,22 @@ PlayerSetting::PlayerSetting(const jellyfin::MediaSource& src) {
         });
 
 /// Fullscreen
-#if defined(__linux__) || defined(_WIN32)
+#if defined(__APPLE__) || defined(__linux__) || defined(_WIN32)
     btnFullscreen->init(
         "main/setting/others/fullscreen"_i18n, conf.getItem(AppConfig::FULLSCREEN, false), [](bool value) {
+            VideoContext::FULLSCREEN = value;
             AppConfig::instance().setItem(AppConfig::FULLSCREEN, value);
-            // 设置当前状态
             brls::Application::getPlatform()->getVideoContext()->fullScreen(value);
+        });
+
+    btnAlwaysOnTop->init(
+        "main/setting/others/always_on_top"_i18n, conf.getItem(AppConfig::ALWAYS_ON_TOP, false), [](bool value) {
+            AppConfig::instance().setItem(AppConfig::ALWAYS_ON_TOP, value);
+            brls::Application::getPlatform()->setWindowAlwaysOnTop(value);
         });
 #else
     btnFullscreen->setVisibility(brls::Visibility::GONE);
+    btnAlwaysOnTop->setVisibility(brls::Visibility::GONE);
 #endif
 
     btnBottomBar->init(
@@ -120,6 +127,124 @@ PlayerSetting::PlayerSetting(const jellyfin::MediaSource& src) {
             MPVCore::OSD_ON_TOGGLE = value;
             conf.setItem(AppConfig::OSD_ON_TOGGLE, value);
         });
+
+    /// Player mirror
+    btnVideoMirror->init("main/setting/filter/mirror"_i18n,
+        {
+            "hints/off"_i18n,
+            "main/setting/filter/hflip"_i18n,
+            "main/setting/filter/vflip"_i18n,
+        },
+        MPVCore::VIDEO_FILTER, [&mpv](int value) {
+            MPVCore::VIDEO_FILTER = value;
+            switch (value) {
+            case 1:
+                mpv.command("set", "vf", "hflip");
+                break;
+            case 2:
+                mpv.command("set", "vf", "vflip");
+                break;
+            default:
+                mpv.command("set", "vf", "");
+            }
+            // 如果正在使用硬解，那么将硬解更新为 auto-copy，避免直接硬解因为不经过 cpu 处理导致镜像翻转无效
+            if (MPVCore::HARDWARE_DEC) {
+                const char* hwdec = value > 0 ? "auto-copy" : MPVCore::PLAYER_HWDEC_METHOD.c_str();
+                mpv.command("set", "hwdec", hwdec);
+                brls::Logger::info("MPV hardware decode: {}", hwdec);
+            }
+        });
+
+    /// Player aspect
+    btnVideoAspect->init("main/setting/aspect/header"_i18n,
+        {
+            "main/setting/aspect/auto"_i18n,
+            "main/setting/aspect/stretch"_i18n,
+            "main/setting/aspect/crop"_i18n,
+            "4:3",
+            "16:9",
+        },
+        conf.getOptionIndex(AppConfig::PLAYER_ASPECT), [&mpv, &conf](int value) {
+            auto& opt = conf.getOptions(AppConfig::PLAYER_ASPECT);
+            MPVCore::VIDEO_ASPECT = opt.options.at(value);
+            mpv.setAspect(MPVCore::VIDEO_ASPECT);
+            conf.setItem(AppConfig::PLAYER_ASPECT, MPVCore::VIDEO_ASPECT);
+        });
+
+    /// Subsync
+    double subDelay = mpv.getDouble("sub-delay");
+    btnSubsync->title->setMarginRight(0);
+    btnSubsync->slider->setMarginRight(0);
+    btnSubsync->slider->setPointerSize(20);
+    btnSubsync->setDetailText(fmt::format("{:.1f}", subDelay));
+    btnSubsync->init("main/setting/playback/subsync"_i18n, (subDelay + 10) * 0.05f, [this](float value) {
+        float data = value * 20 - 10.f;
+        MPVCore::instance().setDouble("sub-delay", data);
+        btnSubsync->setDetailText(fmt::format("{:.1f}", data));
+    });
+
+    btnEqualizerReset->registerClickAction([this](View* view) {
+        btnEqualizerBrightness->slider->setProgress(0.5f);
+        btnEqualizerContrast->slider->setProgress(0.5f);
+        btnEqualizerSaturation->slider->setProgress(0.5f);
+        btnEqualizerGamma->slider->setProgress(0.5f);
+        btnEqualizerHue->slider->setProgress(0.5f);
+        return true;
+    });
+    registerHideBackground(btnEqualizerReset);
+    setupEqualizer(btnEqualizerBrightness, "main/setting/equalizer/brightness"_i18n, Equalizer::BRIGHTNESS,
+        mpv.getDouble("brightness"));
+    setupEqualizer(
+        btnEqualizerContrast, "main/setting/equalizer/contrast"_i18n, Equalizer::CONTRAST, mpv.getDouble("contrast"));
+    setupEqualizer(btnEqualizerSaturation, "main/setting/equalizer/saturation"_i18n, Equalizer::SATURATION,
+        mpv.getDouble("saturation"));
+    setupEqualizer(btnEqualizerGamma, "main/setting/equalizer/gamma"_i18n, Equalizer::GAMMA, mpv.getDouble("hue"));
+    setupEqualizer(btnEqualizerHue, "main/setting/equalizer/hue"_i18n, Equalizer::HUE, mpv.getDouble("gamma"));
 }
 
 PlayerSetting::~PlayerSetting() { brls::Logger::debug("PlayerSetting: delete"); }
+
+void PlayerSetting::setupEqualizer(brls::SliderCell* cell, const std::string& title, Equalizer item, double initValue) {
+    if (initValue < -100)
+        initValue = -100;
+    else if (initValue > 100)
+        initValue = 100;
+
+    cell->detail->setWidth(50);
+    cell->title->setWidth(116);
+    cell->title->setMarginRight(0);
+    cell->slider->setStep(0.05f);
+    cell->slider->setMarginRight(0);
+    cell->slider->setPointerSize(20);
+    cell->setDetailText(fmt::format("{:.0f}", initValue));
+    cell->init(title, (initValue + 100) * 0.005f, [cell, item](float value) {
+        auto& mpv = MPVCore::instance();
+        int data = (int)(value * 200 - 100);
+        cell->setDetailText(std::to_string(data));
+        switch (item) {
+        case Equalizer::BRIGHTNESS:
+            mpv.setInt("brightness", data);
+            break;
+        case Equalizer::CONTRAST:
+            mpv.setInt("contrast", data);
+            break;
+        case Equalizer::SATURATION:
+            mpv.setInt("saturation", data);
+            break;
+        case Equalizer::GAMMA:
+            mpv.setInt("hue", data);
+            break;
+        case Equalizer::HUE:
+            mpv.setInt("gamma", data);
+            break;
+        default:;
+        }
+    });
+    registerHideBackground(cell->getDefaultFocus());
+}
+
+void PlayerSetting::registerHideBackground(brls::View* view) {
+    view->getFocusEvent()->subscribe([this](...) { this->setBackgroundColor(nvgRGBAf(0.0f, 0.0f, 0.0f, 0.0f)); });
+    view->getFocusLostEvent()->subscribe(
+        [this](...) { this->setBackgroundColor(brls::Application::getTheme().getColor("brls/backdrop")); });
+}